@@ -18,11 +18,11 @@
         }
     ],
     "require": {
-        "php": ">=7.1.0",
-        "psr/container": "^1.0"
+        "php": ">=8.0",
+        "psr/container": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "6.5.14"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
