@@ -3,7 +3,7 @@
 namespace ConstanzeStandard\Container;
 
 use AbstractTest;
-use ConstanzeStandard\Container\Interfaces\EntryInterface;
+use ConstanzeStandard\Container\Interfaces\EntityInterface;
 
 require_once __DIR__ . '/AbstractTest.php';
 
@@ -11,11 +11,11 @@ class EntryCollectionTest extends AbstractTest
 {
     public function testGetWithId()
     {
-        /** @var EntryInterface $entry */
-        $entry = $this->createMock(EntryInterface::class);
+        /** @var EntityInterface $entry */
+        $entry = $this->createMock(EntityInterface::class);
         $entry->expects($this->once())->method('resolve')->willReturn(10);
-        $entry->expects($this->once())->method('getIdentifier')->willReturn('id');
-        $entryCollection = new EntryCollection();
+        $entry->expects($this->once())->method('getId')->willReturn('id');
+        $entryCollection = new EntityCollection();
         $entryCollection->add($entry);
         $result = $entryCollection->get('id');
         $this->assertEquals($result, 10);
@@ -26,22 +26,22 @@ class EntryCollectionTest extends AbstractTest
      */
     public function testGetWithoutId()
     {
-        /** @var EntryInterface $entry */
-        $entry = $this->createMock(EntryInterface::class);
+        /** @var EntityInterface $entry */
+        $entry = $this->createMock(EntityInterface::class);
         $entry->expects($this->exactly(0))->method('resolve')->willReturn(10);
-        $entry->expects($this->once())->method('getIdentifier')->willReturn('id');
-        $entryCollection = new EntryCollection();
+        $entry->expects($this->once())->method('getId')->willReturn('id');
+        $entryCollection = new EntityCollection();
         $entryCollection->add($entry);
         $entryCollection->get('nothing');
     }
 
     public function testGetWithIdWithAlias()
     {
-        /** @var EntryInterface $entry */
-        $entry = $this->createMock(EntryInterface::class);
+        /** @var EntityInterface $entry */
+        $entry = $this->createMock(EntityInterface::class);
         $entry->expects($this->once())->method('resolve')->willReturn(10);
-        $entry->expects($this->once())->method('getIdentifier')->willReturn('id');
-        $entryCollection = new EntryCollection();
+        $entry->expects($this->once())->method('getId')->willReturn('id');
+        $entryCollection = new EntityCollection();
         $entryCollection->add($entry);
         $entryCollection->alias('newId', 'id');
         $result = $entryCollection->get('newId');
@@ -53,11 +53,11 @@ class EntryCollectionTest extends AbstractTest
      */
     public function testRemoveWithId()
     {
-        /** @var EntryInterface $entry */
-        $entry = $this->createMock(EntryInterface::class);
+        /** @var EntityInterface $entry */
+        $entry = $this->createMock(EntityInterface::class);
         $entry->expects($this->exactly(0))->method('resolve')->willReturn(10);
-        $entry->expects($this->once())->method('getIdentifier')->willReturn('id');
-        $entryCollection = new EntryCollection();
+        $entry->expects($this->once())->method('getId')->willReturn('id');
+        $entryCollection = new EntityCollection();
         $entryCollection->add($entry);
         $entryCollection->remove('id');
         $entryCollection->get('id');
@@ -65,11 +65,11 @@ class EntryCollectionTest extends AbstractTest
 
     public function testResolveWithConstructValue()
     {
-        /** @var EntryInterface $entry */
-        $entry = $this->createMock(EntryInterface::class);
+        /** @var EntityInterface $entry */
+        $entry = $this->createMock(EntityInterface::class);
         $entry->expects($this->once())->method('resolve')->with([1, 2], true)->willReturn(10);
-        $entry->expects($this->once())->method('getIdentifier')->willReturn('id');
-        $entryCollection = new EntryCollection([$entry]);
+        $entry->expects($this->once())->method('getId')->willReturn('id');
+        $entryCollection = new EntityCollection([$entry]);
         $result = $entryCollection->resolve('id', [1, 2], true);
         $this->assertEquals($result, 10);
     }
